use explicit dtypes when computing the loss
Loss functions used to cast everything to float() and assume it worked.
This behaviour changed in PyTorch 2.13 which doesn't silently tolerate
this loosey-goosey behaviour anymore. Hooray for minor releases with
backwards incompatible changes!

Bump version to 0.23.9

+++ src/matchescu/matching/matchers/ml/deeper/training/_dataset.py
@@ -57,7 +57,7 @@ class DeepERDataset(MatchescuDataset):
         left_enc, right_enc = to_deeper_repr(
             left, right, self.__tokenizer, self.__attr_map, self.__max_len
         )
-        labels = torch.tensor(self._labels[idx], dtype=torch.float)
+        labels = torch.tensor(self._labels[idx], dtype=torch.int64)
         return left_enc, right_enc, labels
 
     def _collate(
